nm_join_bayes: let caller override data's dependent variable
nm_join_bayes() assumes that the dependent variable from the data is
"DV", but this won't be the case if the NONMEM model remapped another
column to DV with, say, `$INPUT ... FOO=DV ...`.

Note that the npde::autonpde() call is the only spot that needs to
know the DV column name.  It will show up in the tables, but the logic
in prep_nm_join_data() discards the DV column without using its name.

Thanks to @timwaterhouse for pointing out this issue.