nm_join_bayes: don't hard code DV when reading tables
When prep_nm_join_data() reads in the bbr-bayes-join.tab files from
the chain submodels, it assumes that they have a DV column.  This
won't be the case if the NONMEM model remapped another data column.
For example

  $INPUT [...] DV1=DV [...]

will lead to DV1 in the table files.

The next commit will add a dv_col that lets the caller specify a
remapped DV name for use with npde::autonpde(), so we could also
update prep_nm_join_data() to accept that argument.  However, that
seems like an overkill given that prep_nm_join_data() is just
filtering out the value.

Instead follow the approach used in the origin run_sims()
implementation from the example project, and drop extra table columns
that collide with data columns.  Note that prep_nm_join_data() already
checks that `data` does _not_ have collisions with the table columns
that we use downstream (e.g., "PRED", "EWRES").