install_torsten: tighten cleanup of temporary .tar.gz file
install_torsten() downloads the tar.gz to the installation directory
and handles the cleanup with a file.remove() call triggered by an
on.exit().  That on.exit(), however, happens after the download.file()
and untar() calls; if any errors that happen between when the file is
created and the on.exit(), the file won't be removed.

Get dest_file with withr::local_tempfile().  In addition to not having
to worry about the cleanup ourselves, this has the advantage that,
even if the cleanup fails, the file is under a temporary directory
rather than the installation directory.