install_torsten: drop unnecessary condition
The conditions for the version/release_url handling end with an else
block that's entered if version is NULL and release_url is NULL.  This
else block then has another 'is version NULL?' check.  Based on the
upstream handling, we know version is NULL at this point.

Drop the condition to avoid having an unreachable code path where
download_url isn't set.