install_torsten tests: always use quiet=TRUE
The install_torsten() tests emit output to the console despite being
surrounded by expect_output() and expect_message().  The uncaught
output is coming via non-R library code sending messages to stderr.

We could use a type=message sink to get this output, but it's not
worth the effort for an argument that's relayed as is to
cmdstanr::check_cmdstan_toolchain(), cmdstanr::rebuild_cmdstan(), and
utils::download.file().  Instead always pass `quiet = TRUE`.