install_torsten: take version over release_url as intended
The caller shouldn't specify both the version and release_url, but,
when they do, install_torsten() is supposed to take the version.
However, the release_url condition is handled regardless of whether
the upstream 'not null version' arm is entered, so the release_url
actually takes precedence [*].

Only execute the 'not null release_url' arm if the 'not null version'
arm hasn't been executed.

[*] This issue is flagged by a failing test:
    https://github-drone.metrumrg.com/metrumresearchgroup/bbr.bayes/337/1/4