test-nm-join-bayes.R: increase n_post for tests
The npde 'not positive definite' error with low n_post values is
expected due to the input data (in particular given number of
observations per subjects).

The npde test is the main thing that this matters for, but let's just
bump all the nm_join_bayes calls to use a higher value because that
only adds ~2s to the test run.  (Note that the test model uses only 50
iterations.)

(This is very unlikely to be a concern in any real setting, so there's
not much point in guarding.)

Thanks to Tim W for assessing the cause of the error.