shrinkage: simplify setup of a test
To check the error case where a shk file doesn't have a TYPE=4, the
tests overwrite the shk files with a custom data frame in CSV format.
That's confusing because it relies on the reader knowing that the
internal details of how the file is read means that it's okay for it
to be a CSV and for the header to be on a different line the real shk
files.

Instead make the minimal tweak to the existing shk file, replacing the
TYPE=4 value.