nmbayes: allow specifying name when getting chain files
chain_paths() expects the name to be {id}-{chain} (i.e. the name of
the submodel).  That works for many files, including those that are
automatically named based on the name of the model, but users may want
to grab files that don't follow the above naming scheme (e.g., custom
tables).

Add a `name` argument that enables grabbing files with arbitrary
names.