test-copy-model-from.R: adjust check for bbr change
copy_model_as_nmbayes() calls bbr::copy_model_from() underneath.
copy_model_from() is supposed to overwrite the $PROBLEM record for the
copied model, but it only handled upper cases record names until bbr's
13fb8e5f (copy_control_stream: use nmrec instead of regex,
2023-07-21).

A copy_model_as_nmbayes() test checks for lines that it expects to be
in the parent model but not the new one.  It uses "$prob" for the
record name, and expects the old bbr behavior (i.e. the same problem
line is in both the parent and child), leading to a failure with the
unreleased bbr.

Use an upper case $PROB in the parent model so that copy_model_from()
updates the record for bbr's versions before and after 13fb8e5f.

Re: https://github.com/metrumresearchgroup/bbr/pull/604