inst/model/nonmem: ignore .msf files
The nonmem/bayes/1100 model specifies a 1100.msf output file via
$EST's MSFO option.  bbi adds a pattern to the chain submodel's
.gitignore for {ID}.msf (i.e. 1100-{chain}.msf) [1].  If not for a bug
in update_model_id() [2], the submodel control stream would point to
1100-{chain}.msf, so the bbi rule would be in effect.

The next commit will drop the update_model_id() call on the submodel,
so the MSFO option for each submodel control stream will point to
1100.msf, which won't be covered by the bbi-generated pattern.

The core issue is that bbr.bayes is breaking bbi's expectations here.
It's not clear to me how many other cases there are like this and what
(if anything) should be done about it on either the bbr.bayes or bbi
side.

For now add a rule ignoring all *.msf entries under inst/model/nonmem
to prevent tracking these in this repo.

[1] There are also entries for related pattern with the same stem,
    such as 1100-1_ETAS.msf.

[2] ... as is, the MFSO option is mangled and no msf file is written.

    https://github.com/metrumresearchgroup/bbr/issues/611