nm_join_bayes: support .superset=TRUE
The goal is for nm_join_bayes() to support all of the arguments that
bbr::nm_join() does, but .superset=TRUE requires extra handling
because passing the extra rows to mrgsolve will trigger an error.

We can identify the extra rows by looking for rows that have NAs for
all non-data columns.  Filter those out before simulating and then
join the result back to the full set of data rows before returning.

Closes #100.