nm_join_bayes: use future for epred and ipred simulations
The main decision here is how to batch the samples (i.e. what will be
an element for an iteration of a future.apply [1] map function).  The
extremes are

 * use each sample/row as an element.  The main downsides of that are
   1) it involves a join operation for each row and 2) future will
   pre-generate a seed for each row, which adds overhead.

 * use all samples as an element, defeating the point of going through
   future

Go in between by batching the samples so that at least 200 samples [2]
are in each element sent to future.  Callers will likely want to tune
this in some cases for better performance given their model and other
details (e.g. future configuration), so expose an argument for
overriding the value.

Importantly choosing the number of batches by looking at the result of
future::nbrOfWorkers() is not a good approach because it means the
result is only reproducible if executed in an environment where
nbrOfWorkers() gives the same answer.

[1] I initially tried using furrr rather than future.apply given that
    purrr is used widely in this code base.  However, running the
    furrr-backed code sequentially attaches purrr to the current
    session.  Rather than investigating what's triggering that, I
    switched to using future.apply.  (Aside from the suggested
    dependency, the details aren't exposed to the user, so the choice
    doesn't matter too much.)

[2] That value isn't backed by testing.  Ideally we'd pick a value
    based on testing some realistic scenarios.  And, if we're going to
    change the default, it be good to change it sooner than later to
    avoid changing the result of nm_join_bayes() calls in the wild.