model_summary: signal an error for stan and nmbayes models
model_summary.bbi_stan_model() and model_summary.bbi_nmbayes_model()
both relay the result of read_fit_model(), with the plan to eventually
implement a tailored summary (gh-27).  The idea is that, while those
objects are not summaries, they are _entry points_ to summarizing the
results.  However, these objects, especially the posterior arrays for
nmbayes models, are not something we want in summary_log().

Avoid that problem by updating these methods to error instead of warn.
Given that 1) this was simply relaying the read_fit_model() result
with a warning to use that instead and 2) the early stage of the
bbr.bayes package, there is very low risk of this breaking any scripts
in the wild (and, if it does, the fix is obvious).

Re: https://github.com/metrumresearchgroup/bbr/pull/645#discussion_r1473479118