stan submit_model: add .bbi_args for generic/method consistency
In a8e18fa (submit_model.bbi_stan_model: drop unsupported args,
2023-01-16), I dropped .bbi_args from the submit_model() method for
Stan models.  Under newer versions of R, the check rightly flags this:

    checking S3 generic/method consistency ... WARNING
    [...]
    submit_model:
      function(.mod, .bbi_args, .mode, ..., .overwrite, .config_path,
               .wait, .dry_run)
    submit_model.bbi_stan_model:
      function(.mod, .mode, ..., .overwrite)

Add back .bbi_args so that the meaning of the second positional
argument is consistent between the generic and Stan submit_model()
methods.

Note that this _could_ break some existing callers that pass `.mode`
as a positional argument, but it's very unlikely that anyone is doing
that given 1) the early stage of this package and 2) that Stan models
do not currently support any other .mode value than the "local"
default.