install_torsten: rework version matching to account for substrings
install_torsten() checks if the user-specified version is in the set
of known releases by grepping for the specified string.  The
motivation for using grep was to accommodate Torsten's inconsistent
tag naming [1].  Here's the current tag list returned by
get_torsten_download_url:

  * v0.91.0
  * v0.91.0rc3
  * v0.91.0rc2
  * 0.91.0rc1
  * torsten_v0.90.0
  * torsten_v0.89.1
  * v0.89.0

With the grep approach, the caller can specify "0.89.1" and match
torsten_v0.89.1 without needing to know whether the tag had a "v"or
"torsten_v" prefix.

This approach fails, however, if a version is a substring of another.
For example, there is no way to select v0.91.0 with the current
approach: specifying "v0.91.0" will also match v0.91.0rc2 and
v0.91.0rc3.

Instead accept "torsten_v{version}", "torsten_{version}" [2],
"v{version}", and "{version}" as valid ways to specify the _complete_
version.  So "0.91.0" would match v0.91.0, and "0.91.0rc2" would match
v0.91.0rc2.

As with the previous approach, this depends on Torsten not creating
multiple tags for any given version, and an error is still raised if a
version matches multiple tags.

This means that providing a partial version, for example "9.0" to
select v0.89.0, no longer works.  However, that behavior was
undocumented and not the original motivation of using grep.

[1] https://github.com/metrumresearchgroup/bbr.bayes/pull/95#discussion_r1281038479

[2] No existing tags use this variant, but it's a plausible enough to
    be worth including for future-proofing.