test-install-torsten.R: install v0.91.0 for Ubuntu 24.04 compatibility
The Torsten installation is failing in the ubuntu-latest (currently
Ubuntu 24.04) workflows due to a failure to compile the example model.

The actual error isn't surfaced, but in an ubuntu:noble Docker image
running `make build` from Torsten v0.89.1's cmdstan subdirectory fails
with

  ../tbb_2020.3/include/tbb/task.h:300:20: error: declaration of
  'tbb::task& tbb::internal::task_prefix::task()' changes meaning of
  'task'

The build for Torsten 0.90.0 also fails, but the build for Torsten
v0.91.0 succeeds.

That failure doesn't seem to align with the CI's because the latter is
happening with the example model build _after_ the
cmdstanr::cmdstan_make_local call.  But, in any case, updating to
Torsten v0.91.0 also resolves the CI failure for ubuntu-latest.