nmbayes submit_model: ensure inits are generated synchronously
submit_model.bbi_nmbayes_model does an upfront submission of a model
to generate the init.chn file that's used as input for the main
sampling models.

The caller-specified .wait value is propagated to the initial model
submission.  This creates a race between init.chn being created and
the main sampling runs reading it in.

Do the initial submission under .wait=TRUE.

Fixes #155