ci: error if check emits note if running under latest R
Ideally the check job would execute under error="note" so that a CI
error would flag new issues at the note level (such as the Rd target
issues fixed by the previous commit).  There hasn't been a
straightforward way to do that in the context of bbr.bayes, though,
because the check always has two notes (one related to the number of
imports and another related to the package size).

With R 4.5, the import and size messages are now marked as "info"
instead of a note, so run the release build with error="note".